# cli/openapi/agent-openapi.json
{
  "components": {
    "schemas": {
...
      "BulkMarkReadRequest": {
        "description": "Request to mark multiple content items as read.",
        "example": {
          "content_ids": [
            123,
            456,
            789
          ]
        },
        "properties": {
          "content_ids": {
            "description": "List of content IDs to mark as read",
            "items": {
              "type": "integer"
            },
            "minItems": 1,
            "title": "Content Ids",
            "type": "array"
          }
        },
        "required": [
          "content_ids"
        ],
        "title": "BulkMarkReadRequest",
        "type": "object"
      },
...
      "ContentListResponse": {
        "description": "Response for content list endpoint.",
        "example": {
          "available_dates": [
            "2025-06-19",
            "2025-06-18"
          ],
          "content_types": [
            "article",
            "podcast",
            "news"
          ],
          "contents": [
            {
              "classification": "to_read",
              "content_type": "article",
              "created_at": "2025-06-19T10:30:00Z",
              "id": 123,
              "platform": "substack",
              "processed_at": "2025-06-19T10:35:00Z",
              "short_summary": "This article explores...",
              "source": "Tech Blog",
              "status": "completed",
              "title": "Understanding AI in 2025",
              "url": "https://example.com/article"
            }
          ],
          "meta": {
            "has_more": true,
            "next_cursor": "eyJsYXN0X2lkIjoxMjN9",
            "page_size": 25,
            "total": 1
          }
        },
        "properties": {
          "available_dates": {
            "description": "List of available dates (YYYY-MM-DD)",
            "items": {
              "type": "string"
            },
            "title": "Available Dates",
            "type": "array"
          },
          "content_types": {
            "description": "Available content types for filtering",
            "items": {
              "$ref": "#/components/schemas/ContentType"
            },
            "title": "Content Types",
            "type": "array"
          },
          "contents": {
            "description": "List of content items",
            "items": {
              "$ref": "#/components/schemas/ContentSummaryResponse"
            },
            "title": "Contents",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetadata",
            "description": "Pagination metadata for the response"
          }
        },
        "required": [
          "contents",
          "available_dates",
          "content_types",
          "meta"
        ],
        "title": "ContentListResponse",
        "type": "object"
      },
...
      "ContentType": {
        "description": "Supported content types in storage and API contracts.",
        "enum": [
          "article",
          "podcast",
          "news",
          "insight_report",
          "unknown"
        ],
        "title": "ContentType",
        "type": "string"
      },
      "ConvertNewsItemResponse": {
        "description": "Response for converting a news item into long-form article content.",
        "properties": {
          "already_exists": {
            "title": "Already Exists",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "new_content_id": {
            "title": "New Content Id",
            "type": "integer"
          },
          "news_item_id": {
            "title": "News Item Id",
            "type": "integer"
          },
          "status": {
            "default": "success",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "news_item_id",
          "new_content_id",
          "already_exists",
          "message"
        ],
        "title": "ConvertNewsItemResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginationMetadata": {
        "description": "Pagination metadata for responses.",
        "properties": {
          "has_more": {
            "default": false,
            "description": "Whether more results are available",
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "description": "Opaque cursor token for next page (null if no more results)",
            "nullable": true,
            "title": "Next Cursor",
            "type": "string"
          },
          "page_size": {
            "description": "Number of items in current response",
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total number of items (expensive to compute, may be omitted)",
            "nullable": true,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page_size"
        ],
        "title": "PaginationMetadata",
        "type": "object"
      },
...
      "ScraperConfigResponse": {
        "description": "Response model for scraper configs.",
        "example": {
          "config": {
            "feed_url": "https://example.substack.com/feed"
          },
          "created_at": "2025-06-24T12:00:00Z",
          "display_name": "Import AI",
          "feed_url": "https://example.substack.com/feed",
          "id": 1,
          "is_active": true,
          "limit": 10,
          "scraper_type": "substack",
          "stats": {
            "average_interval_hours": 24.0,
            "completed_count": 7,
            "interval_sample_size": 3,
            "latest_processed_at": "2026-03-30T21:30:00Z",
            "latest_publication_at": "2026-03-30T20:00:00Z",
            "next_expected_at": "2026-04-01T08:00:00Z",
            "processing_count": 1,
            "total_count": 8,
            "unread_count": 3
          }
        },
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "nullable": true,
            "title": "Display Name",
            "type": "string"
          },
          "feed_url": {
            "nullable": true,
            "title": "Feed Url",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "limit": {
            "nullable": true,
            "title": "Limit",
            "type": "integer"
          },
          "scraper_type": {
            "title": "Scraper Type",
            "type": "string"
          },
          "stats": {
            "$ref": "#/components/schemas/ScraperConfigStatsResponse",
            "nullable": true
          }
        },
        "required": [
          "id",
          "scraper_type",
          "config",
          "is_active",
          "created_at"
        ],
        "title": "ScraperConfigResponse",
        "type": "object"
      },
      "ScraperConfigStatsResponse": {
        "description": "Derived stats for a single scraper configuration.",
        "properties": {
          "average_interval_hours": {
            "nullable": true,
            "title": "Average Interval Hours",
            "type": "number"
          },
          "completed_count": {
            "title": "Completed Count",
            "type": "integer"
          },
          "interval_sample_size": {
            "default": 0,
            "title": "Interval Sample Size",
            "type": "integer"
          },
          "latest_processed_at": {
            "format": "date-time",
            "nullable": true,
            "title": "Latest Processed At",
            "type": "string"
          },
          "latest_publication_at": {
            "format": "date-time",
            "nullable": true,
            "title": "Latest Publication At",
            "type": "string"
          },
          "next_expected_at": {
            "format": "date-time",
            "nullable": true,
            "title": "Next Expected At",
            "type": "string"
          },
          "processing_count": {
            "title": "Processing Count",
            "type": "integer"
          },
          "total_count": {
            "title": "Total Count",
            "type": "integer"
          },
          "unread_count": {
            "title": "Unread Count",
            "type": "integer"
          }
        },
        "required": [
          "total_count",
          "completed_count",
          "unread_count",
          "processing_count"
        ],
        "title": "ScraperConfigStatsResponse",
        "type": "object"
      },
...
      "SubscribeToFeedRequest": {
        "description": "Request to subscribe to a detected feed.",
        "example": {
          "display_name": "Example Newsletter",
          "feed_type": "substack",
          "feed_url": "https://example.substack.com/feed"
        },
        "properties": {
          "display_name": {
            "nullable": true,
            "title": "Display Name",
            "type": "string"
          },
          "feed_type": {
            "title": "Feed Type",
            "type": "string"
          },
          "feed_url": {
            "title": "Feed Url",
            "type": "string"
          }
        },
        "required": [
          "feed_url",
          "feed_type"
        ],
        "title": "SubscribeToFeedRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Filtered machine-oriented API contract for the standalone newsly-agent CLI.",
    "title": "Newsly Agent CLI API",
    "version": "2.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/news/items": {
      "get": {
        "description": "Return the visible representative news feed for the current user.",
        "operationId": "listNewsItems",
        "parameters": [
          {
            "description": "Filter by read status",
            "in": "query",
            "name": "read_filter",
            "required": false,
            "schema": {
              "default": "unread",
              "description": "Filter by read status",
              "pattern": "^(all|read|unread)$",
              "title": "Read Filter",
              "type": "string"
            }
          },
          {
            "description": "Opaque cursor token",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "description": "Opaque cursor token",
              "nullable": true,
              "title": "Cursor",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List visible news items",
        "tags": [
          "news"
        ]
      }
    },
    "/api/news/items/mark-read": {
      "post": {
        "description": "Mark the given visible representative news items as read.",
        "operationId": "markNewsItemsRead",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkMarkReadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Marknewsitemsread",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark visible news items as read",
        "tags": [
          "news"
        ]
      }
    },
    "/api/news/items/{news_item_id}": {
      "get": {
        "description": "Return one visible representative news item.",
        "operationId": "getNewsItem",
        "parameters": [
          {
            "in": "path",
            "name": "news_item_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": true,
              "minimum": 0,
              "title": "News Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get one news item",
        "tags": [
          "news"
        ]
      }
    },
    "/api/news/items/{news_item_id}/convert-to-article": {
      "post": {
        "description": "Convert one representative news item into saved article content.",
        "operationId": "convertNewsItemToArticle",
        "parameters": [
          {
            "in": "path",
            "name": "news_item_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": true,
              "minimum": 0,
              "title": "News Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConvertNewsItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Convert one news item into article content",
        "tags": [
          "news"
        ]
      }
    },
    "/api/scrapers/": {
      "get": {
        "description": "List scraper configurations for the current user.",
        "operationId": "listScraperConfigs",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "nullable": true,
              "title": "Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "types",
            "required": false,
            "schema": {
              "nullable": true,
              "title": "Types",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScraperConfigResponse"
                  },
                  "title": "Response Listscraperconfigs",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Scraper Configs",
        "tags": [
          "sources"
        ]
      }
    },
    "/api/scrapers/subscribe": {
      "post": {
        "description": "Subscribe to a feed detected from content.\n\nConvenience endpoint that creates a scraper config from a detected feed.",
        "operationId": "subscribeScrapersToFeed",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeToFeedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScraperConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Subscribe To Feed",
        "tags": [
          "sources"
        ]
      }
    }
  },
...
}